+++ test/e2e/common/secrets.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e
+package common
 
 import (
 	"fmt"
@@ -82,10 +82,10 @@ var _ = framework.KubeDescribe("Secrets", func() {
 			},
 		}
 
-		framework.TestContainerOutput("consume secrets", f.Client, pod, 0, []string{
+		f.TestContainerOutput("consume secrets", pod, 0, []string{
 			"content of file \"/etc/secret-volume/data-1\": value-1",
 			"mode of file \"/etc/secret-volume/data-1\": -rw-r--r--",
-		}, f.Namespace.Name)
+		})
 	})
 
 	It("should be consumable in multiple volumes in a pod [Conformance]", func() {
